Intent: modularizar operadores de jitter y remesh Operators involved: [Self-organization, Transition] Affected invariants: [#4, #7] Key changes: - Extraer caché de jitter y utilidades a submódulo dedicado - Mover funciones de remallado a nuevo submódulo - Simplificar operators.__init__ y reexportar API Expected risks/dissonances: posibles errores de importación; mitigados con reexportes y pruebas Metrics: C(t), Si sin cambios; νf y phase estabilizadas Equivalence map: tnfr.operators.* -> tnfr.operators.remesh/jitter
